# src/backend/laravel-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API",
        "description": "Documentation de l'API de votre application.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reglement/findAll": {
            "get": {
                "tags": [
                    "ReglementInterieur"
                ],
                "summary": "Get all reglementInterieur",
                "description": "Retrieve a list of all reglementInterieur",
                "operationId": "findAllreglementInterieurs",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "JWT token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer {your_token}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "reglements": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ReglementInterieur"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error - Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reglement/create": {
            "post": {
                "tags": [
                    "ReglementInterieur"
                ],
                "summary": "Create a new reglementInterieur",
                "description": "Create a new reglementInterieur resource",
                "operationId": "createreglementInterieur",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "JWT token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "Bearer {your_token}"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libelle"
                                ],
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "example": "Etre ponctuel"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Error - Invalid request data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Could not create this "
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "object",
                                            "example": {
                                                "libelle": [
                                                    "The libelle field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ReglementInterieur created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ReglementInterieur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reglement/findOne/{id}": {
            "get": {
                "tags": [
                    "ReglementInterieur"
                ],
                "summary": "Get reglement information",
                "description": "Get information about a specific reglementInterieur",
                "operationId": "showOneReglementInterieur",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "JWT token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "Bearer {your_token}"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of reglementInterieur to get information for",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Error - Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error - Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "reglementInterieur not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ReglementInterieur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reglement/update": {
            "post": {
                "tags": [
                    "ReglementInterieur"
                ],
                "summary": "Update a reglementInterieur's information",
                "description": "Update a reglementInterieur's information",
                "operationId": "updatereglementInterieur",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "JWT token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "Bearer {your_token}"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libelle"
                                ],
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Etre assidue"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Error - Invalid request data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "example": {
                                                "libelle": [
                                                    "The libelle field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error - Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error - Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "ReglementInterieur not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "reglement": {
                                            "$ref": "#/components/schemas/ReglementInterieur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reglement/delete/{id}": {
            "delete": {
                "tags": [
                    "ReglementInterieur"
                ],
                "summary": "Delete a reglementInterieur",
                "description": "Delete a reglementInterieur resource",
                "operationId": "deleteReglementInterieur",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "JWT token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "Bearer {your_token}"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of reglementInterieur to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Error - Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error - Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "ReglementInterieur not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ReglementInterieur": {
                "required": [
                    "libelle"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": "1"
                    },
                    "role": {
                        "description": "Un nouveau reglement interieur",
                        "type": "string",
                        "readOnly": true
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "deleted_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/deleted_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ReglementInterieur"
                }
            },
...
        }
    }
}